Rework library to work on arduino-cli v0.18
