+++ library.json
@@ -1,0 +1,18 @@
+{
+    "name": "Crc16",
+    "keywords": "CRC, crc, crc16",
+    "description": "Arduino Library for CRC-16 checks",
+    "authors": [
+        {
+            "name": "Vincenzo Mennella",
+            "maintainer": true
+        }
+    ],
+    "repository": {
+        "type": "git",
+        "url": "https://https://github.com/vinmenn/Crc16.git"
+    },
+    "version": "0.1.2",
+    "frameworks": "arduino",
+    "platforms": "*"
+}
