+++ library.properties
@@ -1,0 +1,11 @@
+name=Crc16
+version=0.1.2
+author=Vincenzo Mennella
+maintainer=Hugo Arganda <hugo.arganda@gmail.com>
+sentence=CRC16 Library
+paragraph=CRC16 Library
+category=Data Processing
+url=https://github.com/vinmenn/Crc16.git
+architectures=*
+includes=Crc16.h
+depends=
